refactor(libs): ディレクトリ構造を変更
