+++ src/libs/scan.ts
